Move NewUrlControls into ClientTests
