+++ react-app/src/components/ClientTests/ModalNewUrl/NewUrl.js
@@ -1,0 +1,8 @@
+export default function NewUrl({ client, newUrl }) {
+  return (
+    <>
+      <h1>Send test to {client.code}</h1>
+      <p>{newUrl}</p>
+    </>
+  );
+}
