+++ react-app/src/components/ClientTests/ModalNewUrl/index.js
@@ -2,14 +2,19 @@
 import NewUrl from './NewUrl';
 
 // import context
-import { Modal } from '../../../../../context/Modal';
+import { Modal } from '../../../context/Modal';
 
-export default function ModalNewUrl({ showModal, setShowModal, newUrl }) {
+export default function ModalNewUrl({
+  showModal,
+  setShowModal,
+  client,
+  newUrl,
+}) {
   return (
     <>
       {showModal && (
         <Modal onClose={() => setShowModal(false)}>
-          <NewUrl newUrl={newUrl} setShowModal={setShowModal} />
+          <NewUrl newUrl={newUrl} client={client} />
         </Modal>
       )}
     </>
