+++ react-app/src/components/ClientTests/index.js
@@ -1,6 +1,9 @@
 import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
+// import components
+import NewUrlControls from './NewUrlControls';
+
 export default function ClientTests() {
   const { clientId } = useParams();
   const client = useSelector((state) => state.clients[clientId]);
@@ -11,6 +14,7 @@ export default function ClientTests() {
     <>
       <h1>{client.code}</h1>
       <p>test</p>
+      <NewUrlControls client={client} />
     </>
   );
 }
